requirements.txt
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
python-dateutil==2.8.2
icalendar>=5.0.0
gpt4all>=2.5.0
pytesseract>=0.3.10
Pillow>=10.0.0

# WhatsApp Chat Exporter dependencies (for WhatsApp database import)
jinja2>=3.0.0
bleach>=6.0.0
markupsafe>=2.0.0

# Database dependencies
psycopg2-binary>=2.9.0  # PostgreSQL/Supabase support
